change: initial commit - refactoring of code to accommodate multiple DB providers
