+++ BeGeneric.Backend.Abstractions/Builder/IBeGenericBackendBuilder.cs
@@ -1,0 +1,11 @@
+﻿using BeGeneric.Backend.Common.Models;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace BeGeneric.Backend.Common.Builder;
+
+public interface IBeGenericBackendBuilder
+{
+    IServiceCollection Services { get; }
+    BeConfiguration Configuration { get; }
+    Type PrimaryKeyType { get; }
+}
