+++ BeGeneric.Backend.Abstractions/Models/BeConfiguration.cs
@@ -1,0 +1,7 @@
+﻿namespace BeGeneric.Backend.Common.Models;
+
+public class BeConfiguration
+{
+    public List<ColumnMetadataDefinition> Metadata { get; set; }
+    public List<EntityDefinition> Entities { get; set; }
+}
