+++ BeGeneric.Backend.MsSql/BeGeneric.Backend.Database.MsSql.csproj
@@ -1,0 +1,36 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+    <PackageId>BeGeneric.Backend.Database.MsSql</PackageId>
+    <Version>0.0.0.1-alpha</Version>
+    <Authors>Be-Generic</Authors>
+    <Company>Be-Generic</Company>
+    <PackageDescription>Addon package for Be-Generic to enable usage of MS SQL.</PackageDescription>
+    <RepositoryUrl>https://github.com/be-generic/be-generic</RepositoryUrl>
+    <PackageLicenseFile>LICENSE.md</PackageLicenseFile>
+    <PackageReleaseNotes>
+      Initial release of the MS SQL database provider.
+    </PackageReleaseNotes>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Data.SqlClient" Version="6.0.2" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.2" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\BeGeneric.Backend.Database\BeGeneric.Backend.Database.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Update="LICENSE.md">
+      <PackagePath></PackagePath>
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+      <Pack>true</Pack>
+    </None>
+  </ItemGroup>
+
+</Project>
