+++ BeGeneric.Backend.Sample/BeGeneric.Backend.Sample.csproj
@@ -14,6 +14,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\BeGeneric.Backend.MsSql\BeGeneric.Backend.Database.MsSql.csproj" />
     <ProjectReference Include="..\BeGeneric.Backend\BeGeneric.Backend.csproj" />
   </ItemGroup>
