+++ BeGeneric.Backend/BeGeneric.Backend.csproj
@@ -26,6 +26,11 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\BeGeneric.Backend.Abstractions\BeGeneric.Backend.Common.csproj" />
+    <ProjectReference Include="..\BeGeneric.Backend.Database\BeGeneric.Backend.Database.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
     <None Update="LICENSE.md" Pack="true" PackagePath="">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
