+++ BeGeneric.Backend/Services/GenericBackend/DataRequestObject.cs
@@ -1,4 +1,7 @@
-﻿namespace BeGeneric.Backend.Services.GenericBackend
+﻿using BeGeneric.Backend.Common.Models;
+using BeGeneric.Backend.Database;
+
+namespace BeGeneric.Backend.Services.GenericBackend
 {
     public class DataRequestObject : ComparerObject
     {
