+++ BeGeneric.Backend/Services/GenericBackend/Helpers/DTOObjectHelper.cs
@@ -1,6 +1,8 @@
-﻿using BeGeneric.Backend.Settings;
-using System.Reflection.Emit;
+﻿using System.Reflection.Emit;
 using System.Reflection;
+using BeGeneric.Backend.Common;
+using BeGeneric.Backend.Common.Helpers;
+using BeGeneric.Backend.Common.Models;
 
 namespace BeGeneric.Backend.Services.GenericBackend.Helpers
 {
