--- BeGeneric.Backend/Settings/BeConfiguration.cs
@@ -1,9 +1,0 @@
-﻿namespace BeGeneric.Backend.Settings
-{
-    public class BeConfiguration
-    {
-        public List<ColumnMetadataDefinition> Metadata { get; set; }
-        public List<EntityDefinition> Entities { get; set; }
-    }
-
-}
