--- BeGeneric.Backend/Settings/EntityRoleDefinition.cs
@@ -1,16 +1,0 @@
-﻿namespace BeGeneric.Backend.Settings
-{
-    public class EntityRoleDefinition
-    {
-        public bool GetOne { get; set; } = false;
-        public bool GetAll { get; set; } = false;
-        public bool Post { get; set; } = false;
-        public bool Put { get; set; } = false;
-        public bool Delete { get; set; } = false;
-
-        public string? ViewFilter { get; set; }
-        public string? EditFilter { get; set; }
-
-        public string RoleKey { get; set; } = string.Empty;
-    }
-}
